break welcome section into parts
