+++ src/lib/env.ts
@@ -10,7 +10,7 @@ export const env = createEnv({
     S3_ACCESS_KEY_ID: z.string().min(1),
     S3_SECRET_ACCESS_KEY: z.string().min(1),
     S3_REGION: z.string().min(1),
-    NEXT_PUBLIC_CF_URL: z.string().min(1),
+    // S3_CF_PUBLIC_ENDPOINT: z.string().min(1),
     API_KEY: z.string().min(1),
   },
   client: {
