--- src/stories/WelcomeSectieon.stories.tsx
@@ -1,20 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react'
-
-import { WelcomeSection } from '@/components/WelcomeSection'
-
-const meta = {
-  title: 'WelcomeSection',
-  component: WelcomeSection,
-  parameters: {
-    layout: 'fullscreen',
-  },
-
-  tags: ['autodocs'],
-} satisfies Meta<typeof WelcomeSection>
-
-export default meta
-type Story = StoryObj<typeof meta>
-
-export const Primary: Story = {
-  parameters: {},
-}
