+++ src/stories/WelcomeSection.stories.ts
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react'
 import { fn } from '@storybook/test'
 
-import { WelcomeSection } from '@/components/WelcomeSection'
+import { WelcomeSection } from '@/app/(frontend)/_components/sections/HomePage/Welcome/WelcomeSection'
 
 const meta = {
   title: 'WelcomeSection',
